refactor: create manageWord folder for word management related scripts
