+++ src/manageKeys/IManageKeys.ts
@@ -1,11 +1,8 @@
 import { Game } from "../Game.js";
-
 export interface IManageKeys {
     word: string;
     game: Game;
 
-    getRandomWord(): Promise<string>;
-    printPickedWord(word: string): void;
     setupGameWithListeners(word: string): void;
     setupClickListeners(game: Game): void;
     setupKeyDownListener(game: Game): void;
