+++ src/manageWords/Word.ts
