+++ src/manageWords/printPickedWord.ts
@@ -1,0 +1,3 @@
+export function printPickedWord(word: string): void {
+    console.log(word);
+}
