+++ src/script.ts
@@ -1,12 +1,12 @@
 import { ManageKeys } from "./manageKeys/ManageKeys.js";
+import { getRandomWord } from "./manageWords/getRandomWord.js";
+import { printPickedWord } from "./manageWords/printPickedWord.js";
 
-// Función principal para iniciar el juego
 async function startGame(): Promise<void> {
     const manageKeysInstance = new ManageKeys();
-    const pickedWord: string = await manageKeysInstance.getRandomWord();
-    manageKeysInstance.printPickedWord(pickedWord);
+    const pickedWord: string = await getRandomWord();
+    printPickedWord(pickedWord);
     manageKeysInstance.setupGameWithListeners(pickedWord);
 }
 
-// Iniciar el juego
 startGame();
